List of Plant Species Name accordingly with the Leipzig Catalogue of Plants (LCP)

The 'LCPspecies_table' contains 1.314.778 records belonging to all the vascular plant taxa names checked and listed from the "Leipzig Catalogue of Plants" (LCP).

Format

A data frame with 1314778 observations on the following 7 variables.

Input.Taxon

A character vector. This variable contains the list of all 1.314.778 vascular plant taxa names listed by the authors of the Leipzig Catalogue of Plants.

Status

A character vector. This variable provides a description if a taxon is classified as 'valid', 'synonym', 'unresolved' or 'external'. The 'unresolved' rank means that the status of the plant name could be either valid or synonym, but the information available does not allow a definitive decision (see the main text Freiberg et al. for more details).

PL.comparison

A character vector. This field provides a direct comparison with 'The Plant List' (TPL; The Plant List http://www.theplantlist.org/ accessed: 1.1. 2013) reporting further information such as 'identical', 'synonym', 'other synonym', 'different authors', 'missing', 'misspelling', 'unresolved'.

PL.alternative

A character vector. This field provides a possible alternative name suggested by TPL.

Output.Taxon

A character vector. This variable contains the list of the accepted plant taxa names (350.766) according to the LCP.

Family

A character vector. It provides the corresponding family name of the 'Input.Taxon', staying empty if the Status is unresolved.

Order

A character vector. The Order column provides the corresponding order name of the 'Input.Taxon', staying empty if the Status is unresolved.

Source

https://idata.idiv.de/ddm/Data/ShowData/1806

References

The Leipzig Catalogue of Plants (LCP) - An improved taxonomic reference list for all known vascular plants.

Examples

data(LCPspecies_table)
str(LCPspecies_table)

#> 'data.frame':	1502 obs. of  7 variables:
#>  $ Input.Taxon   : chr  "Hibiscus abelmoschus L." "Hibiscus abelmoschus var. betulifolius Mast." "Hibiscus abelmoschus var. genuinus Hochr." "Hibiscus abutiloides Willd." ...
#>  $ Status        : chr  "synonym" "synonym" "synonym" "synonym" ...
#>  $ PL.comparison : chr  "identical" "identical" "identical" "identical" ...
#>  $ PL.alternative: chr  "" "" "" "" ...
#>  $ Output.Taxon  : chr  "Abelmoschus moschatus Medik." "Abelmoschus moschatus Medik." "Abelmoschus moschatus Medik." "Talipariti tiliaceum (L.) Fryxell" ...
#>  $ Family        : chr  "Malvaceae" "Malvaceae" "Malvaceae" "Malvaceae" ...
#>  $ Order         : chr  "Malvales" "Malvales" "Malvales" "Malvales" ...